backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
psycopg2-binary>=2.9.9
Pillow>=10.1.0
python-dotenv>=1.0.0
requests>=2.31.0
django-modeltranslation>=0.18.11

